Remove std type trait equivalents
The original intent of the type_traits.h file was to provide custom
type traits, and low-complexity equivalents to std type traits. This
was mainly to improve compile times over including type_traits.

Unfortunately as this file grew and the needs of the engine gew the
file started to get complex. Plus, some of these traits on GCC were
*very* complex as they didn't provide built-ins like other compilers.

For these reasons I decided to change this file to just expose
<type_traits> and a few custom traits.