Prefix getters in schema codegen with Get*
This fixes an issue where return types sometimes match the function
name. Previously did some hacky things to work around it on MSVC
but GCC was not pleased. Just always prefixing has some benefits in
terms of reducing codegen complexity, being more compatible, and
also make it more clear when you are reading a value vs calling
a getter on the base StructReader family of classes.