Improve code generation build ordering
The schemac, bin2c, and shaderc tools will now create new modules that
perform the code generation. The module that listed the options for
running these tools now depends on the generated modules. This ensures
that the generated code is always created first before trying to compile
any module that needs it.

- Add a prepare handler for module keys which runs before project generation.
- Fix build failures when generated files are missing.
- Move display attributes into the editor project.